refactor(Event, Config): 환경문제로 미 인식된 파일 추가, Test용 DB설정 파일 위치 변경
