+++ common/src/main/java/kitchenpos/common/event/UngroupOrderTableEvent.java
@@ -1,0 +1,21 @@
+package kitchenpos.common.event;
+
+import java.util.List;
+
+public class UngroupOrderTableEvent {
+    private final Long tableGroupId;
+    private final List<Long> orderTableIds;
+    
+    public UngroupOrderTableEvent(Long tableGroupId, List<Long> orderTableIds) {
+        this.tableGroupId = tableGroupId;
+        this.orderTableIds = orderTableIds;
+    }
+
+    public Long getTableGroupId() {
+        return tableGroupId;
+    }
+
+    public List<Long> getOrderTableIds() {
+        return this.orderTableIds;
+    }
+}
