+++ common/src/main/java/kitchenpos/common/event/ValidateEmptyTableEvent.java
@@ -1,0 +1,13 @@
+package kitchenpos.common.event;
+
+public class ValidateEmptyTableEvent {
+    private Long orderTableId;
+
+    public ValidateEmptyTableEvent(Long orderTableId) {
+        this.orderTableId = orderTableId;
+    }
+
+    public Long getOrderTableId() {
+        return orderTableId;
+    }
+}
