+++ table/src/test/java/kitchenpos/table/presentation/TableRestControllerTest.java
@@ -12,7 +12,7 @@ import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import kitchenpos.table.dto.OrderTableDto;
-import kitchenpos.testassistance.config.TestConfig;
+import kitchenpos.common.testassistance.config.TestConfig;
 
 @DisplayName("주문테이블 API기능에 관한")
 public class TableRestControllerTest extends TestConfig {
